// src/app/interfaces/forms-juridica-contratacion.ts
export interface FormsJuridicaContratacion {
  id: string;
  tipo: string,
  numSolicitud: string;
  fecha: string;
  tipoAdquisicion: string;
  otraAdquisicion:string;
  cargoA: string;
  nombreConvenio?: string;
  nombreUnidad?: string;
  responAreaFinan: string;
  nombreResponAreaFinan: string;
  nombreSolicitante: string;
  VBAreaJuridica: string;
}